+++ src/snowflake/snowpark/_internal/analyzer/snowflake_plan.py
@@ -30,6 +30,7 @@ from snowflake.snowpark._internal.analyzer.query_plan_analysis_utils import (
 from snowflake.snowpark._internal.analyzer.table_function import (
     GeneratorTableFunction,
     TableFunctionRelation,
+    TableFunctionJoin,
 )
 
 if TYPE_CHECKING:
@@ -1324,6 +1325,44 @@ class SnowflakePlanBuilder:
             source_plan,
         )
 
+    def find_table_function_in_sql_tree(self, plan: SnowflakePlan):
+        from snowflake.snowpark._internal.analyzer.select_statement import (
+            SelectTableFunction,
+            Selectable,
+        )
+
+        if isinstance(plan, SelectTableFunction) and isinstance(
+            plan.snowflake_plan.source_plan, TableFunctionJoin
+        ):
+            if (
+                plan.snowflake_plan.source_plan.right_cols == ["*"]
+                and len(plan.snowflake_plan.children_plan_nodes) == 1
+            ):
+                child_plan = plan.snowflake_plan.children_plan_nodes[0]
+                if isinstance(child_plan, Selectable):
+                    child_plan = child_plan.snowflake_plan
+                plan.snowflake_plan.source_plan.right_cols = (
+                    plan.snowflake_plan.quoted_identifiers[
+                        len(child_plan.quoted_identifiers) :
+                    ]
+                )
+                new_plan = self.session._analyzer.resolve(
+                    plan.snowflake_plan.source_plan
+                )
+                plan._snowflake_plan = new_plan
+                return plan
+        plan_2_resolve = None
+        for node in plan.children_plan_nodes:
+            plan_2_resolve = (
+                self.find_table_function_in_sql_tree(node) or plan_2_resolve  # type: ignore
+            )
+        if plan_2_resolve:
+            return self.session._analyzer.resolve(
+                plan.snowflake_plan.source_plan  # type: ignore
+                if isinstance(plan, Selectable)
+                else plan.source_plan
+            )
+
     def create_or_replace_dynamic_table(
         self,
         name: str,
@@ -1341,6 +1380,13 @@ class SnowflakePlanBuilder:
         source_plan: Optional[LogicalPlan],
         iceberg_config: Optional[dict] = None,
     ) -> SnowflakePlan:
+        child_find_table_function = copy.deepcopy(child)
+        child_find_table_function = self.find_table_function_in_sql_tree(
+            child_find_table_function
+        )
+        if child_find_table_function is not None:
+            child = child_find_table_function
+
         if len(child.queries) != 1:
             raise SnowparkClientExceptionMessages.PLAN_CREATE_DYNAMIC_TABLE_FROM_DDL_DML_OPERATIONS()